New set of command line programs: AnnotateVariantsKraken: annotate VCF with kraken2 classifications ExtractFragmentsToFastq.java: exports named reads to paired fastq UnmappedSequencesToFastq.java: exports unmapped sequences to fastq InsertedSequencesToFasta.java: exports single breakend/breakpoint inserted sequences to fasta ExtractBestSequencesBasedOnReport: extracts the sequences in a given taxonomy with the most hits SubsetToTaxonomy: subsets kraken2 output those in given taxonomy

+++ src/main/java/au/edu/wehi/idsv/vcf/VcfInfoAttributes.java
@@ -78,7 +78,7 @@ public enum VcfInfoAttributes {
 	INSERTED_SEQUENCE_REPEATMASKER_REPEAT_CLASS("INSRMRC", 1, VCFHeaderLineType.String, "Inserted sequence repeatmasker repeat class."),
 	INSERTED_SEQUENCE_REPEATMASKER_ORIENTATION("INSRMRO", 1, VCFHeaderLineType.String, "Inserted sequence repeatmasker repeat orientation."),
 	INSERTED_SEQUENCE_REPEATMASKER_OVERLAP("INSRMP", 1, VCFHeaderLineType.Float, "Portion of inserted sequence whose alignment overlaps the repeatmasker repeat. 1.0 indicates the inserted sequence entirely mapping to the repeat."),
-	INSERTED_SEQUENCE_NCBI_TAXONOMY_ID("INSTAXID", 1, VCFHeaderLineType.Float, "NCBI Taxonomy identifier for inserted sequence.");
+	INSERTED_SEQUENCE_NCBI_TAXONOMY_ID("INSTAXID", 1, VCFHeaderLineType.Integer, "NCBI Taxonomy identifier for inserted sequence.");
 
 	private final VCFInfoHeaderLine header;
 	private final String tag;
